cloud/api-gateway/pyproject.toml
[project]
name = "seld-api-gateway"
version = "0.1.0"
description = "SELD Digital Twin — FastAPI REST/WebSocket gateway"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.18",
    "redis>=5.2.0",
    "structlog>=24.4.0",
    "seld-common",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.hatch.build.targets.wheel]
packages = ["src/api_gateway"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[tool.uv.sources]
seld-common = { path = "../../libs/seld-common", editable = true }
